refactor: expose project vision tools
