+++ packages/cli/src/commands/mcp.ts
@@ -11,7 +11,7 @@ import {
   getProjectBasicAuthCredentials,
   type BuilderNamespace,
 } from "@webstudio-is/project-build/contracts";
-import { diffPngFiles } from "@webstudio-is/project-build/visual";
+import { diffPngFiles } from "@webstudio-is/project-build/vision";
 import {
   publicApiOperationRequiresServerSupport,
   publicApiOperations,
