+++ packages/project-build/src/entrypoints.test.ts
@@ -20,7 +20,7 @@ describe("package entrypoints", () => {
   test("does not expose Node-only surfaces through generic conditions", () => {
     for (const entrypoint of [
       "./mcp",
-      "./visual",
+      "./vision",
       "./transfer/server",
       "./server",
     ]) {
